move coverages tests to own file
