+++ tests/test_api.py
@@ -45,7 +45,6 @@ from pygeoapi.api import (
     API, APIRequest, FORMAT_TYPES, validate_bbox, validate_datetime,
     validate_subset, F_HTML, F_JSON, F_JSONLD, F_GZIP, __version__, apply_gzip,
 )
-from pygeoapi.api.coverages import get_collection_coverage
 from pygeoapi.api.itemtypes import (
     get_collection_schema, get_collection_queryables, get_collection_item,
     get_collection_items, manage_collection_item)
@@ -691,18 +690,6 @@ def test_describe_collections(config, api_):
     assert rsp_headers['Content-Type'] == FORMAT_TYPES[F_HTML]
     assert rsp_headers['Content-Language'] == 'en-US'
 
-    req = mock_request()
-    rsp_headers, code, response = api_.describe_collections(req,
-                                                            'gdps-temperature')
-    collection = json.loads(response)
-
-    assert collection['id'] == 'gdps-temperature'
-    assert len(collection['links']) == 10
-    assert collection['extent']['spatial']['grid'][0]['cellsCount'] == 2400
-    assert collection['extent']['spatial']['grid'][0]['resolution'] == 0.15000000000000002  # noqa
-    assert collection['extent']['spatial']['grid'][1]['cellsCount'] == 1201
-    assert collection['extent']['spatial']['grid'][1]['resolution'] == 0.15
-
     # hiearchical collections
     rsp_headers, code, response = api_.describe_collections(
         req, 'naturalearth/lakes')
@@ -753,16 +740,6 @@ def test_get_collection_schema(config, api_):
 
     assert 'properties' in schema
     assert len(schema['properties']) == 5
-
-    req = mock_api_request({'f': 'json'})
-    rsp_headers, code, response = get_collection_schema(
-        api_, req, 'gdps-temperature')
-    assert rsp_headers['Content-Type'] == 'application/schema+json'
-    schema = json.loads(response)
-
-    assert 'properties' in schema
-    assert len(schema['properties']) == 1
-    assert schema['properties']['1']['type'] == 'number'
 
 
 def test_get_collection_queryables(config, api_):
@@ -1402,109 +1379,6 @@ def test_get_collection_item_json_ld(config, api_):
     rsp_headers, code, response = get_collection_item(api_, req, 'obs', '371')
     assert rsp_headers['Content-Type'] == FORMAT_TYPES[F_JSONLD]
     assert rsp_headers['Content-Language'] == 'fr-CA'
-
-
-def test_get_collection_coverage(config, api_):
-    req = mock_api_request()
-    rsp_headers, code, response = get_collection_coverage(
-        api_, req, 'obs')
-
-    assert code == HTTPStatus.BAD_REQUEST
-
-    req = mock_api_request({'properties': '12'})
-    rsp_headers, code, response = get_collection_coverage(
-        api_, req, 'gdps-temperature')
-
-    assert code == HTTPStatus.BAD_REQUEST
-
-    req = mock_api_request({'subset': 'bad_axis(10:20)'})
-    rsp_headers, code, response = get_collection_coverage(
-        api_, req, 'gdps-temperature')
-
-    assert code == HTTPStatus.BAD_REQUEST
-
-    req = mock_api_request({'f': 'blah'})
-    rsp_headers, code, response = get_collection_coverage(
-        api_, req, 'gdps-temperature')
-
-    assert code == HTTPStatus.BAD_REQUEST
-
-    req = mock_api_request({'f': 'html'})
-    rsp_headers, code, response = get_collection_coverage(
-        api_, req, 'gdps-temperature')
-
-    assert code == HTTPStatus.BAD_REQUEST
-    assert rsp_headers['Content-Type'] == 'text/html'
-
-    req = mock_api_request(HTTP_ACCEPT='text/html')
-    rsp_headers, code, response = get_collection_coverage(
-        api_, req, 'gdps-temperature')
-
-    # NOTE: This test used to assert the code to be 200 OK,
-    #       but it requested HTML, which is not available,
-    #       so it should be 400 Bad Request
-    assert code == HTTPStatus.BAD_REQUEST
-    assert rsp_headers['Content-Type'] == 'text/html'
-
-    req = mock_api_request({'subset': 'Lat(5:10),Long(5:10)'})
-    rsp_headers, code, response = get_collection_coverage(
-        api_, req, 'gdps-temperature')
-
-    assert code == HTTPStatus.OK
-    content = json.loads(response)
-
-    assert content['domain']['axes']['x']['num'] == 35
-    assert content['domain']['axes']['y']['num'] == 35
-    assert 'TMP' in content['parameters']
-    assert 'TMP' in content['ranges']
-    assert content['ranges']['TMP']['axisNames'] == ['y', 'x']
-
-    req = mock_api_request({'bbox': '-79,45,-75,49'})
-    rsp_headers, code, response = get_collection_coverage(
-        api_, req, 'gdps-temperature')
-
-    assert code == HTTPStatus.OK
-    content = json.loads(response)
-
-    assert content['domain']['axes']['x']['start'] == -79.0
-    assert content['domain']['axes']['x']['stop'] == -75.0
-    assert content['domain']['axes']['y']['start'] == 49.0
-    assert content['domain']['axes']['y']['stop'] == 45.0
-
-    req = mock_api_request({
-        'subset': 'Lat(5:10),Long(5:10)',
-        'f': 'GRIB'
-    })
-    rsp_headers, code, response = get_collection_coverage(
-        api_, req, 'gdps-temperature')
-
-    assert code == HTTPStatus.OK
-    assert isinstance(response, bytes)
-
-    req = mock_api_request(HTTP_ACCEPT='application/x-netcdf')
-    rsp_headers, code, response = get_collection_coverage(
-        api_, req, 'cmip5')
-
-    assert code == HTTPStatus.OK
-    assert rsp_headers['Content-Type'] == 'application/x-netcdf'
-
-    # req = mock_api_request({
-    #     'subset': 'time("2006-07-01T06:00:00":"2007-07-01T06:00:00")'
-    # })
-    # rsp_headers, code, response = get_collection_coverage(api_, req, 'cmip5')
-    #
-    # assert code == HTTPStatus.OK
-    # assert isinstance(json.loads(response), dict)
-
-    # req = mock_api_request({'subset': 'lat(1:2'})
-    # rsp_headers, code, response = get_collection_coverage(api_, req, 'cmip5')
-    #
-    # assert code == HTTPStatus.BAD_REQUEST
-    #
-    # req = mock_api_request({'subset': 'lat(1:2)'})
-    # rsp_headers, code, response = get_collection_coverage(api_ req, 'cmip5')
-    #
-    # assert code == HTTPStatus.NO_CONTENT
 
 
 def test_validate_bbox():
